Revert "feat(terraform): TF Parser  - Move funcs and consts to utils file (#4550)"
This reverts commit 6bf81c7d07c1f27defba94e83a93998bb9e140ae.
