--- checkov/terraform/modules/__init__.py
@@ -1,2 +1,0 @@
-from checkov.terraform.modules.module_utils import *  # noqa
-from checkov.terraform.modules.module_objects import *  # noqa
