+++ checkov/terraform/parser.py
@@ -1,27 +1,49 @@
 from __future__ import annotations
 
+import json
 import logging
 import os
+import re
+from collections.abc import Sequence
+from collections import defaultdict
 from copy import deepcopy
-from typing import Optional, Dict, Mapping, Set, Tuple, Callable, Any, List
+from json import dumps, loads
+from pathlib import Path
+from typing import Optional, Dict, Mapping, Set, Tuple, Callable, Any, List, TYPE_CHECKING
+import itertools
 
 import deep_merge
+import hcl2
+from lark import Tree
 
 from checkov.common.runners.base_runner import filter_ignored_paths, IGNORE_HIDDEN_DIRECTORY_ENV, strtobool
 from checkov.common.util.consts import DEFAULT_EXTERNAL_MODULES_DIR, RESOLVED_MODULE_ENTRY_NAME
+from checkov.common.util.json_utils import CustomJSONEncoder
 from checkov.common.util.type_forcers import force_list
 from checkov.common.variables.context import EvaluationContext
+from checkov.terraform.checks.utils.dependency_path_handler import unify_dependency_path
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
 from checkov.terraform.graph_builder.graph_components.module import Module
+from checkov.terraform.graph_builder.utils import remove_module_dependency_in_path
 from checkov.terraform.module_loading.content import ModuleContent
 from checkov.terraform.module_loading.module_finder import load_tf_modules
 from checkov.terraform.module_loading.registry import module_loader_registry as default_ml_registry, \
     ModuleLoaderRegistry
-from checkov.common.util.parser_utils import get_tf_definition_key_from_module_dependency, \
-    TERRAFORM_NESTED_MODULE_PATH_ENDING, is_acceptable_module_param
-from checkov.terraform.modules.module_utils import load_or_die_quietly, safe_index, \
-    remove_module_dependency_from_path, get_module_dependency_map, get_module_dependency_map_support_nested_modules, \
-    clean_parser_types, serialize_definitions, get_new_module
+from checkov.common.util.parser_utils import eval_string, find_var_blocks, is_nested, \
+    get_tf_definition_key_from_module_dependency, \
+    get_module_from_full_path, get_abs_path, TERRAFORM_NESTED_MODULE_PATH_ENDING, TERRAFORM_NESTED_MODULE_PATH_PREFIX
+
+if TYPE_CHECKING:
+    from typing_extensions import TypeAlias
+
+
+_Hcl2Payload: TypeAlias = "dict[str, list[dict[str, Any]]]"
+
+external_modules_download_path = os.environ.get('EXTERNAL_MODULES_DIR', DEFAULT_EXTERNAL_MODULES_DIR)
+GOOD_BLOCK_TYPES = {BlockType.LOCALS, BlockType.TERRAFORM}  # used for cleaning bad tf definitions
+
+ENTITY_NAME_PATTERN = re.compile(r"[^\W0-9][\w-]*")
+RESOLVED_MODULE_PATTERN = re.compile(r"\[.+\#.+\]")
 
 
 def _filter_ignored_paths(root: str, paths: list[str], excluded_paths: list[str] | None) -> None:
@@ -106,10 +128,10 @@ class Parser:
 
     def parse_file(self, file: str, parsing_errors: Optional[Dict[str, Exception]] = None) -> Optional[Dict[str, Any]]:
         if file.endswith(".tf") or file.endswith(".tf.json") or file.endswith(".hcl"):
-            parse_result = load_or_die_quietly(file, parsing_errors)
+            parse_result = _load_or_die_quietly(file, parsing_errors)
             if parse_result:
-                parse_result = serialize_definitions(parse_result)
-                parse_result = clean_parser_types(parse_result)
+                parse_result = self._serialize_definitions(parse_result)
+                parse_result = self._clean_parser_types(parse_result)
                 return parse_result
 
         return None
@@ -262,12 +284,12 @@ class Parser:
             var_value_and_file_map[key[7:]] = value, f"env:{key}"
             self.external_variables_data.append((key[7:], value, f"env:{key}"))
         if hcl_tfvars:  # terraform.tfvars
-            data = load_or_die_quietly(hcl_tfvars, self.out_parsing_errors, clean_definitions=False)
+            data = _load_or_die_quietly(hcl_tfvars, self.out_parsing_errors, clean_definitions=False)
             if data:
-                var_value_and_file_map.update({k: (safe_index(v, 0), hcl_tfvars.path) for k, v in data.items()})
-                self.external_variables_data.extend([(k, safe_index(v, 0), hcl_tfvars.path) for k, v in data.items()])
+                var_value_and_file_map.update({k: (_safe_index(v, 0), hcl_tfvars.path) for k, v in data.items()})
+                self.external_variables_data.extend([(k, _safe_index(v, 0), hcl_tfvars.path) for k, v in data.items()])
         if json_tfvars:  # terraform.tfvars.json
-            data = load_or_die_quietly(json_tfvars, self.out_parsing_errors)
+            data = _load_or_die_quietly(json_tfvars, self.out_parsing_errors)
             if data:
                 var_value_and_file_map.update({k: (v, json_tfvars.path) for k, v in data.items()})
                 self.external_variables_data.extend([(k, v, json_tfvars.path) for k, v in data.items()])
@@ -322,7 +344,7 @@ class Parser:
     def _load_files(self, files: list[os.DirEntry]):
         def _load_file(file: os.DirEntry):
             parsing_errors = {}
-            result = load_or_die_quietly(file, parsing_errors)
+            result = _load_or_die_quietly(file, parsing_errors)
             # the exceptions type can un-pickleable
             for path, e in parsing_errors.items():
                 parsing_errors[path] = e
@@ -430,7 +452,7 @@ class Parser:
                     # Special handling for local sources to make sure we aren't double-parsing
                     if source.startswith("./") or source.startswith("../"):
                         source = os.path.normpath(
-                            os.path.join(os.path.dirname(remove_module_dependency_from_path(file)), source))
+                            os.path.join(os.path.dirname(_remove_module_dependency_in_path(file)), source))
 
                     version = module_call_data.get("version", "latest")
                     if version and isinstance(version, list):
@@ -541,8 +563,8 @@ class Parser:
                              download_external_modules=download_external_modules,
                              external_modules_download_path=external_modules_download_path, excluded_paths=excluded_paths,
                              vars_files=vars_files, external_modules_content_cache=external_modules_content_cache)
-        tf_definitions = clean_parser_types(tf_definitions)
-        tf_definitions = serialize_definitions(tf_definitions)
+        tf_definitions = self._clean_parser_types(tf_definitions)
+        tf_definitions = self._serialize_definitions(tf_definitions)
 
         module = None
         if create_graph:
@@ -580,10 +602,10 @@ class Parser:
         source: str,
     ) -> Tuple[Module, Dict[str, Dict[str, Any]]]:
         if self.enable_nested_modules:
-            module_dependency_map, tf_definitions, dep_index_mapping = get_module_dependency_map_support_nested_modules(tf_definitions)
+            module_dependency_map, tf_definitions, dep_index_mapping = self.get_module_dependency_map_support_nested_modules(tf_definitions)
         else:
-            module_dependency_map, tf_definitions, dep_index_mapping = get_module_dependency_map(tf_definitions)
-        module = get_new_module(
+            module_dependency_map, tf_definitions, dep_index_mapping = self.get_module_dependency_map(tf_definitions)
+        module = self.get_new_module(
             source_dir=source_dir,
             module_dependency_map=module_dependency_map,
             module_address_map=self.module_address_map,
@@ -600,6 +622,36 @@ class Parser:
                     logging.warning(f'Failed to add block {blocks[block_type]}. Error:')
                     logging.warning(e, exc_info=False)
         return module, tf_definitions
+
+    @staticmethod
+    def _clean_parser_types(conf: dict[str, Any]) -> dict[str, Any]:
+        if not conf:
+            return conf
+
+        sorted_keys = list(conf.keys())
+        first_key_type = type(sorted_keys[0])
+        if first_key_type is None:
+            return {}
+
+        if all(isinstance(x, first_key_type) for x in sorted_keys):
+            sorted_keys.sort()
+
+        # Create a new dict where the keys are sorted alphabetically
+        sorted_conf = {key: conf[key] for key in sorted_keys}
+        for attribute, values in sorted_conf.items():
+            if attribute == 'alias':
+                continue
+            if isinstance(values, list):
+                sorted_conf[attribute] = Parser._clean_parser_types_lst(values)
+            elif isinstance(values, dict):
+                sorted_conf[attribute] = Parser._clean_parser_types(values)
+            elif isinstance(values, str) and values in ('true', 'false'):
+                sorted_conf[attribute] = True if values == 'true' else False
+            elif isinstance(values, set):
+                sorted_conf[attribute] = Parser._clean_parser_types_lst(list(values))
+            elif isinstance(values, Tree):
+                sorted_conf[attribute] = str(values)
+        return sorted_conf
 
     def get_file_key_with_nested_data(self, file, nested_data):
         if not nested_data:
@@ -618,6 +670,164 @@ class Parser:
                                                     nested_data.get('nested_modules_data'))
         return get_tf_definition_key_from_module_dependency(key, f"{file}{nested_key}", module_index)
 
+    @staticmethod
+    def _clean_parser_types_lst(values: list[Any]) -> list[Any]:
+        for idx, val in enumerate(values):
+            if isinstance(val, dict):
+                values[idx] = Parser._clean_parser_types(val)
+            elif isinstance(val, list):
+                values[idx] = Parser._clean_parser_types_lst(val)
+            elif isinstance(val, str):
+                if val == 'true':
+                    values[idx] = True
+                elif val == 'false':
+                    values[idx] = False
+            elif isinstance(val, set):
+                values[idx] = Parser._clean_parser_types_lst(list(val))
+        str_values_in_lst = [val for val in values if isinstance(val, str)]
+        str_values_in_lst.sort()
+        result_values = [val for val in values if not isinstance(val, str)]
+        result_values.extend(str_values_in_lst)
+        return result_values
+
+    @staticmethod
+    def _serialize_definitions(tf_definitions: dict[str, _Hcl2Payload]) -> dict[str, _Hcl2Payload]:
+        return loads(dumps(tf_definitions, cls=CustomJSONEncoder))
+
+    @staticmethod
+    def get_next_vertices(evaluated_files: list[str], unevaluated_files: list[str]) -> tuple[list[str], list[str]]:
+        """
+        This function implements a lazy separation of levels for the evaluated files. It receives the evaluated
+        files, and returns 2 lists:
+        1. The next level of files - files from the unevaluated_files which have no unresolved dependency (either
+            no dependency or all dependencies were evaluated).
+        2. unevaluated - files which have yet to be evaluated, and still have pending dependencies
+
+        Let's say we have this dependency tree:
+        a -> b
+        x -> b
+        y -> c
+        z -> b
+        b -> c
+        c -> d
+
+        The first run will return [a, y, x, z] as the next level since all of them have no dependencies
+        The second run with the evaluated being [a, y, x, z] will return [b] as the next level.
+        Please mind that [c] has some resolved dependencies (from y), but has unresolved dependencies from [b].
+        The third run will return [c], and the fourth will return [d].
+        """
+        next_level, unevaluated, do_not_eval_yet = [], [], []
+        for key in unevaluated_files:
+            found = False
+            for eval_key in evaluated_files:
+                if eval_key in key:
+                    found = True
+                    break
+            if not found:
+                do_not_eval_yet.append(key.split(TERRAFORM_NESTED_MODULE_PATH_PREFIX)[0])
+                unevaluated.append(key)
+            else:
+                next_level.append(key)
+
+        move_to_uneval = list(filter(lambda k: k.split(TERRAFORM_NESTED_MODULE_PATH_PREFIX)[0] in do_not_eval_yet, next_level))
+        for k in move_to_uneval:
+            next_level.remove(k)
+            unevaluated.append(k)
+        return next_level, unevaluated
+
+    @staticmethod
+    def get_nested_modules_data_as_list(file_path):
+        path = get_abs_path(file_path)
+        modules_list = []
+
+        while is_nested(file_path):
+            module, index = get_module_from_full_path(file_path)
+            modules_list.append((module, index))
+            file_path = module
+        modules_list.reverse()
+        return modules_list, path
+
+    @staticmethod
+    def get_module_dependency_map_support_nested_modules(tf_definitions):
+        module_dependency_map = defaultdict(list)
+        dep_index_mapping = defaultdict(list)
+        for tf_definition_key in tf_definitions.keys():
+            if not is_nested(tf_definition_key):
+                dir_name = os.path.dirname(tf_definition_key)
+                module_dependency_map[dir_name].append([])
+                continue
+            modules_list, path = Parser.get_nested_modules_data_as_list(tf_definition_key)
+            dir_name = os.path.dirname(path)
+            module_dependency_map[dir_name].append([m for m, i in modules_list])
+            dep_index_mapping[(path, modules_list[-1][0])].append(modules_list[-1][1])
+
+        for key, dir_list in module_dependency_map.items():
+            dir_list.sort()
+            module_dependency_map[key] = list(dir_list for dir_list, _ in itertools.groupby(dir_list))
+        return module_dependency_map, tf_definitions, dep_index_mapping
+
+    @staticmethod
+    def get_module_dependency_map(tf_definitions):
+        """
+        :param tf_definitions, with paths in format 'dir/main.tf[module_dir/main.tf#0]'
+        :return module_dependency_map: mapping between directories and the location of its module definition:
+                {'dir': 'module_dir/main.tf'}
+        :return tf_definitions: with paths in format 'dir/main.tf'
+        """
+        module_dependency_map = {}
+        copy_of_tf_definitions = {}
+        dep_index_mapping: Dict[Tuple[str, str], List[str]] = {}
+        origin_keys = list(filter(lambda k: not k.endswith(TERRAFORM_NESTED_MODULE_PATH_ENDING), tf_definitions.keys()))
+        unevaluated_keys = list(filter(lambda k: k.endswith(TERRAFORM_NESTED_MODULE_PATH_ENDING), tf_definitions.keys()))
+        for file_path in origin_keys:
+            dir_name = os.path.dirname(file_path)
+            module_dependency_map[dir_name] = [[]]
+            copy_of_tf_definitions[file_path] = deepcopy(tf_definitions[file_path])
+
+        next_level, unevaluated_keys = Parser.get_next_vertices(origin_keys, unevaluated_keys)
+        while next_level:
+            for file_path in next_level:
+                path, module_dependency, module_dependency_num = remove_module_dependency_in_path(file_path)
+                dir_name = os.path.dirname(path)
+                current_deps = deepcopy(module_dependency_map[os.path.dirname(module_dependency)])
+                for dep in current_deps:
+                    dep.append(module_dependency)
+                if dir_name not in module_dependency_map:
+                    module_dependency_map[dir_name] = current_deps
+                elif current_deps not in module_dependency_map[dir_name]:
+                    module_dependency_map[dir_name] += current_deps
+                copy_of_tf_definitions[path] = deepcopy(tf_definitions[file_path])
+                origin_keys.append(path)
+                dep_index_mapping.setdefault((path, module_dependency), []).append(module_dependency_num)
+            next_level, unevaluated_keys = Parser.get_next_vertices(origin_keys, unevaluated_keys)
+        for key, dep_trails in module_dependency_map.items():
+            hashes = set()
+            deduped = []
+            for trail in dep_trails:
+                trail_hash = unify_dependency_path(trail)
+                if trail_hash in hashes:
+                    continue
+                hashes.add(trail_hash)
+                deduped.append(trail)
+            module_dependency_map[key] = deduped
+        return module_dependency_map, copy_of_tf_definitions, dep_index_mapping
+
+    @staticmethod
+    def get_new_module(
+            source_dir: str,
+            module_dependency_map: Dict[str, List[List[str]]],
+            module_address_map: Dict[Tuple[str, str], str],
+            external_modules_source_map: Dict[Tuple[str, str], str],
+            dep_index_mapping: Dict[Tuple[str, str], List[str]],
+    ) -> Module:
+        return Module(
+            source_dir=source_dir,
+            module_dependency_map=module_dependency_map,
+            module_address_map=module_address_map,
+            external_modules_source_map=external_modules_source_map,
+            dep_index_mapping=dep_index_mapping
+        )
+
     def add_tfvars(self, module: Module, source: str) -> None:
         if not self.external_variables_data:
             return
@@ -632,3 +842,117 @@ class Parser:
             dirname_path = os.path.dirname(path)
             self.dirname_cache[path] = dirname_path
         return dirname_path
+
+
+def _load_or_die_quietly(
+    file: str | Path, parsing_errors: dict[str, Exception], clean_definitions: bool = True
+) -> _Hcl2Payload | None:
+    """
+Load JSON or HCL, depending on filename.
+    :return: None if the file can't be loaded
+    """
+
+    file_path = os.fspath(file)
+    file_name = os.path.basename(file_path)
+
+    try:
+        logging.debug(f"Parsing {file_path}")
+
+        with open(file_path, "r", encoding="utf-8-sig") as f:
+            if file_name.endswith(".json"):
+                return json.load(f)
+            else:
+                raw_data = hcl2.load(f)
+                non_malformed_definitions = validate_malformed_definitions(raw_data)
+                if clean_definitions:
+                    return clean_bad_definitions(non_malformed_definitions)
+                else:
+                    return non_malformed_definitions
+    except Exception as e:
+        logging.debug(f'failed while parsing file {file_path}', exc_info=True)
+        parsing_errors[file_path] = e
+        return None
+
+
+def _is_valid_block(block: Any) -> bool:
+    if not isinstance(block, dict):
+        return True
+
+    # if the block is empty, there's no need to process it further
+    if not block:
+        return False
+
+    entity_name = next(iter(block.keys()))
+    if re.fullmatch(ENTITY_NAME_PATTERN, entity_name):
+        return True
+    return False
+
+
+def validate_malformed_definitions(raw_data: _Hcl2Payload) -> _Hcl2Payload:
+    return {
+        block_type: [block for block in blocks if _is_valid_block(block)]
+        for block_type, blocks in raw_data.items()
+    }
+
+
+def clean_bad_definitions(tf_definition_list: _Hcl2Payload) -> _Hcl2Payload:
+    return {
+        block_type: [
+            definition
+            for definition in definition_list
+            if block_type in GOOD_BLOCK_TYPES or not isinstance(definition, dict) or len(definition) == 1
+        ]
+        for block_type, definition_list in tf_definition_list.items()
+    }
+
+
+def _to_native_value(value: str) -> Any:
+    if value.startswith('"') or value.startswith("'"):
+        return value[1:-1]
+    else:
+        return eval_string(value)
+
+
+def _remove_module_dependency_in_path(path: str) -> str:
+    """
+    :param path: path that looks like "dir/main.tf[other_dir/x.tf#0]
+    :return: only the outer path: dir/main.tf
+    """
+    if re.findall(RESOLVED_MODULE_PATTERN, path):
+        path = re.sub(RESOLVED_MODULE_PATTERN, '', path)
+    return path
+
+
+def _safe_index(sequence_hopefully: Sequence[Any], index: int) -> Any:
+    try:
+        return sequence_hopefully[index]
+    except IndexError:
+        logging.debug(f'Failed to parse index int ({index}) out of {sequence_hopefully}', exc_info=True)
+        return None
+
+
+def is_acceptable_module_param(value: Any) -> bool:
+    """
+    This function determines if a value should be passed to a module as a parameter. We don't want to pass
+    unresolved var, local or module references because they can't be resolved from the module, so they need
+    to be resolved prior to being passed down.
+    """
+    value_type = type(value)
+    if value_type is dict:
+        for k, v in value.items():
+            if not is_acceptable_module_param(v) or not is_acceptable_module_param(k):
+                return False
+        return True
+    if value_type is set or value_type is list:
+        for v in value:
+            if not is_acceptable_module_param(v):
+                return False
+        return True
+
+    if value_type is not str:
+        return True
+
+    for vbm in find_var_blocks(value):
+        if vbm.is_simple_var():
+            return False
+    return True
