+++ tests/terraform/graph/graph_builder/test_terraform_graph_parser.py
@@ -2,7 +2,6 @@ import os
 
 from lark import Tree
 
-from checkov.terraform.modules.module_utils import clean_parser_types
 from checkov.terraform.parser import Parser
 from unittest import TestCase, mock
 
@@ -12,31 +11,36 @@ TEST_DIRNAME = os.path.dirname(os.path.realpath(__file__))
 class TestParser(TestCase):
     def test_bool_parsing_avoid_remove_non_existing(self):
         conf = {'test': ['Bool'], 'variable': ['aws:SecureTransport'], 'values': [['false']]}
-        actual = clean_parser_types(conf)
+        config_parser = Parser()
+        actual = config_parser._clean_parser_types(conf)
         expected = {'test': ['Bool'], 'variable': ['aws:SecureTransport'], 'values': [[False]]}
         self.assertDictEqual(expected, actual)
 
     def test_bool_parsing_sort_only_lists(self):
         conf = {'enabled_metrics': [['a', 'c', 'b'], 'b', 'a', 'c']}
-        actual = clean_parser_types(conf)
+        config_parser = Parser()
+        actual = config_parser._clean_parser_types(conf)
         expected = {'enabled_metrics': [['a', 'b', 'c'], 'a', 'b', 'c']}
         self.assertDictEqual(expected, actual)
 
     def test_bool_parsing_sort_only_lists_with_bools(self):
         conf = {'enabled_metrics': [['a', 'true', 'false'], 'b', 'true', 'false']}
-        actual = clean_parser_types(conf)
+        config_parser = Parser()
+        actual = config_parser._clean_parser_types(conf)
         expected = {'enabled_metrics': [[True, False, 'a'], True, False, 'b']}
         self.assertDictEqual(expected, actual)
 
     def test_set_parsing_to_list(self):
         conf = {'enabled_metrics': [['a', 'true', 'false'], 'b', 'true', 'false'], 'example_set': [{'1', '2', '3'}]}
-        actual = clean_parser_types(conf)
+        config_parser = Parser()
+        actual = config_parser._clean_parser_types(conf)
         expected = {'enabled_metrics': [[True, False, 'a'], True, False, 'b'], 'example_set': [['1', '2', '3']]}
         self.assertDictEqual(expected, actual)
 
     def test_tree_parsing_to_str(self):
         conf = {'enabled_metrics': [['a', 'true', 'false'], 'b', 'true', 'false'], 'example_set': Tree("data", ["child1", "child2"])}
-        actual = clean_parser_types(conf)
+        config_parser = Parser()
+        actual = config_parser._clean_parser_types(conf)
         expected = {'enabled_metrics': [[True, False, 'a'], True, False, 'b'], 'example_set': 'Tree(\'data\', [\'child1\', \'child2\'])'}
         self.assertDictEqual(expected, actual)
