+++ tests/terraform/parser/test_module.py
@@ -3,8 +3,7 @@ import unittest
 
 import hcl2
 
-from checkov.terraform.modules.module_utils import validate_malformed_definitions, clean_bad_definitions
-from checkov.terraform.parser import Parser
+from checkov.terraform.parser import validate_malformed_definitions, clean_bad_definitions, Parser
 
 
 class ModuleTest(unittest.TestCase):
